Moved "common files" in common dir
