+++ client/src/components/common/SimpleInput.js
