+++ client/src/components/form/Form.js
@@ -2,9 +2,9 @@ import React, {useEffect, useState} from "react";
 import {useDispatch} from "react-redux";
 import useStyles from "./styles";
 import {Paper} from "@material-ui/core";
-import SimpleInput from "./SimpleInput";
-import ImageUploadBox from "./ImageUploadBox";
-import ConfirmOrClearBox from "./ConfrimOrClearBox";
+import SimpleInput from "../common/SimpleInput";
+import ImageUploadBox from "../common/ImageUploadBox";
+import ConfirmOrClearBox from "../common/ConfrimOrClearBox";
 
 const GenericForm = ( {initialState, onSubmit, inputFields} ) => {
     const [data, setData] = useState({...initialState});
@@ -18,7 +18,6 @@ const GenericForm = ( {initialState, onSubmit, inputFields} ) => {
     const clear = () => {
         setData({...initialState});
     };
-
 
     const handleSubmit = async (e) => {
         e.preventDefault();
@@ -41,7 +40,11 @@ const GenericForm = ( {initialState, onSubmit, inputFields} ) => {
             <form id={"genericForm"} autoComplete="off" noValidate className={`${classes.root} ${classes.form}`} onSubmit={handleSubmit}>
                 <div>
                     {inputFields.map((ele) =>
-                        <SimpleInput name={ele.name} label={ele.label} onChange={(e) => handleChangedData(e)} value={data[ele.name]}/>)}
+                        <SimpleInput 
+                            name={ele.name} 
+                            label={ele.label} 
+                            onChange={(e) => handleChangedData(e)} 
+                            value={data[ele.name]}/>)}
                 </div>
                 <ImageUploadBox onUpload={(e) => onUpload(e)}></ImageUploadBox>
                 <ConfirmOrClearBox onClear={clear} onConfirm={handleSubmit} ></ConfirmOrClearBox>
