+++ client/src/components/form/styles.js
@@ -14,10 +14,6 @@ export default makeStyles((theme) => ({
     flexWrap: 'wrap',
     justifyContent: 'center',
   },
-  fileInput: {
-    width: '97%',
-    margin: '10px 0',
-  },
   buttonSubmit: {
     marginBottom: 10,
   },
